Refactor copy-pasted function into separate module
