+++ lib/galaxy/webapps/galaxy/buildapp.py
@@ -19,6 +19,7 @@ import galaxy.model.mapping
 import galaxy.datatypes.registry
 import galaxy.web.framework
 import galaxy.web.framework.webapp
+from galaxy.webapps.util import build_template_error_formatters
 from galaxy import util
 from galaxy.util import asbool
 from galaxy.util.properties import load_app_properties
@@ -713,22 +714,3 @@ def wrap_in_static( app, global_conf, plugin_frameworks=None, **local_conf ):
 
     # URL mapper becomes the root webapp
     return urlmap
-
-
-def build_template_error_formatters():
-    """
-    Build a list of template error formatters for WebError. When an error
-    occurs, WebError pass the exception to each function in this list until
-    one returns a value, which will be displayed on the error page.
-    """
-    formatters = []
-    # Formatter for mako
-    import mako.exceptions
-
-    def mako_html_data( exc_value ):
-        if isinstance( exc_value, ( mako.exceptions.CompileException, mako.exceptions.SyntaxException ) ):
-            return mako.exceptions.html_error_template().render( full=False, css=False )
-        if isinstance( exc_value, AttributeError ) and exc_value.args[0].startswith( "'Undefined' object has no attribute" ):
-            return mako.exceptions.html_error_template().render( full=False, css=False )
-    formatters.append( mako_html_data )
-    return formatters
