+++ lib/galaxy/webapps/util.py
@@ -1,0 +1,17 @@
+def build_template_error_formatters():
+    """
+    Build a list of template error formatters for WebError. When an error
+    occurs, WebError pass the exception to each function in this list until
+    one returns a value, which will be displayed on the error page.
+    """
+    formatters = []
+    # Formatter for mako
+    import mako.exceptions
+
+    def mako_html_data( exc_value ):
+        if isinstance( exc_value, ( mako.exceptions.CompileException, mako.exceptions.SyntaxException ) ):
+            return mako.exceptions.html_error_template().render( full=False, css=False )
+        if isinstance( exc_value, AttributeError ) and exc_value.args[0].startswith( "'Undefined' object has no attribute" ):
+            return mako.exceptions.html_error_template().render( full=False, css=False )
+    formatters.append( mako_html_data )
+    return formatters
